:art: initial refactor of EC PBC
- Moved PBC to its own EC folder.
- Slight refactor of the `execute_processing_block` task
- Added setup.py and modified PBC folder structure to be able
  to build as a pip installable library.
- Published PBC library to pypi as `skasip-pbc==1.2.1`
- Published new Docker image to dockerhub as
  `skasip/processing_block_controller:1.2.1`
