+++ sip/execution_control/processing_block_controller/CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on
 and this project adheres to
  [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2018-12-04
+
+### Changed
+- Updated to skasip-config-db@1.2.1 (which downgrades to redis 2.10.6 from 
+  3.0.1)
+
 ## [1.2.0] - 2018-11-29
 
 ### Changed
