+++ sip/execution_control/processing_block_controller/build_image.sh
@@ -5,7 +5,7 @@ function build_and_publish_image() {
     RED='\033[0;31m'
     BLUE='\033[0;34m'
     NC='\033[0m'
-    VERSION="$(python -c "from release import __version__; print(__version__)")"
+    VERSION="$(python -c "from sip_pbc.release import __version__; print(__version__)")"
     IMAGE=skasip/processing_block_controller
     echo -e "${RED}--------------------------------------------------------------${NC}"
     echo -e "${BLUE}Building and uploading ${IMAGE} image, version = latest, ${VERSION}"
