+++ sip/execution_control/processing_block_controller/sip_pbc/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+"""Processing Block Controller module."""
+from .tasks import APP, version, execute_processing_block
+
+__all__ = [
+    'APP',
+    'version',
+    'execute_processing_block'
+]
