+++ sip/execution_control/processing_block_controller/sip_pbc/release.py
@@ -2,7 +2,7 @@
 """Processing Block Controller release info."""
 __subsystem__ = 'ExecutionControl'
 __service_name__ = 'ProcessingBlockController'
-__version_info__ = (1, 2, 0)
+__version_info__ = (1, 2, 1)
 __version__ = '.'.join(map(str, __version_info__))
 __service_id__ = ':'.join(map(str, (__subsystem__,
                                     __service_name__,
