+++ sip/execution_control/processing_block_controller/tests/__init__.py
@@ -1,0 +1,2 @@
+# coding=utf-8
+"""Tests of the SIP processing block controller."""
