+++ sip/execution_control/processing_block_controller/tests/data/sbi_config_2.json
@@ -19,13 +19,13 @@
         "version": "1.0.1",
         "parameters": {
             "stage1": {
-                "duration": 2
+                "duration": 1
             },
             "stage2": {
-                "duration": 2
+                "duration": 1
             },
             "stage3": {
-                "duration": 2
+                "duration": 1
             }
         }
       }
