+++ sip/execution_control/processing_block_controller/tests/data/workflow_definitions/templates/mock_workflow/1.0.0/stage1/1.0.0/docker-compose.yaml.j2
