+++ sip/execution_control/processing_block_controller/tests/data/workflow_definitions/templates/mock_workflow/1.0.1/stage2/1.0.1/args.json.j2
