+++ sip/execution_control/processing_block_controller/tests/data/workflow_definitions/templates/mock_workflow/1.0.1/stage3/1.0.1/docker-compose.yaml.j2
