+++ sip/execution_control/processing_block_controller/tests/test_processing_block_controller.py
@@ -3,40 +3,39 @@
 
 http://docs.celeryproject.org/en/latest/userguide/testing.html
 """
-
 import json
 from os.path import dirname, join
 
-import celery
-from celery.app.control import Inspect
-
 from sip_config_db import ConfigDb
 from sip_config_db.scheduling import SchedulingBlockInstance
+
+from sip_pbc import APP, execute_processing_block, version
+from sip_pbc.release import __version__
+
 from .test_utils import add_workflow_definitions
-from ..processing_block_controller.release import __version__
-from ..processing_block_controller.tasks import APP, \
-    execute_processing_block, version
 
 DB = ConfigDb()
 
 
 def test_pbc_inspect_tasks():
     """."""
+    from celery.app.control import Inspect
     _inspect = Inspect(app=APP)
-    print('')
-    print('inspect =', _inspect)
-    print('scheduled =', _inspect.scheduled())
-    print('active =', _inspect.active())
-    print('reserved =', _inspect.reserved())
-    print('registered tasks =', _inspect.registered_tasks())
+    # print('')
+    # print('inspect =', _inspect)
+    # print('scheduled =', _inspect.scheduled())
+    # print('active =', _inspect.active())
+    # print('reserved =', _inspect.reserved())
+    # print('registered tasks =', _inspect.registered_tasks())
 
 
 def test_pbc_inspect_workers():
     """."""
-    print('')
-    print('ping workers:', celery.current_app.control.inspect().ping())
-    print('stats', json.dumps(celery.current_app.control.inspect().stats(),
-                              indent=2))
+    # import celery
+    # print('')
+    # print('ping workers:', celery.current_app.control.inspect().ping())
+    # print('stats', json.dumps(celery.current_app.control.inspect().stats(),
+    #                           indent=2))
     # print('workers=', celery.current_app.control.inspect().stats().keys())
 
 
@@ -54,6 +53,8 @@ def test_pbc_execute_workflow():
     test_pbc_execute sip/execution_control/processing_controller
 
     """
+    # import celery
+
     DB.flush_db()
     data_dir = join(dirname(__file__), 'data')
     add_workflow_definitions(join(data_dir, 'workflow_definitions'))
@@ -61,9 +62,8 @@ def test_pbc_execute_workflow():
         sbi_config = json.load(_file)
     sbi = SchedulingBlockInstance.from_config(sbi_config)
 
-    print('NAME:', execute_processing_block.name)
-    state = celery.current_app.events.State()
-    print("STATE:", state)
+    # print('NAME:', execute_processing_block.name)
+    # print("STATE:", celery.current_app.events.State())
 
     pb_ids = sbi.processing_block_ids
     assert len(pb_ids) == 1
@@ -71,4 +71,4 @@ def test_pbc_execute_workflow():
     assert isinstance(pb_ids[0], str)
 
     result = execute_processing_block.delay(pb_ids[0])
-    print('result = ', result)
+    assert result.get() == 'completed'
