--- sip/execution_control/processing_controller/processing_block_controller/README.md
@@ -1,1 +1,0 @@
-# Processing Block Controller
