--- sip/execution_control/processing_controller/processing_block_controller/__init__.py
@@ -1,2 +1,0 @@
-# coding=utf-8
-"""Processing Block Controller module."""
